tests/wpt/tests/html/semantics/forms/form-submission-0/getactionurl.html

<!DOCTYPE html>
<meta charset="utf-8">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>

<iframe src="resources/getactionurl-iframe.html"></iframe>
<script>
async_test(t => {
  window.onmessage = t.step_func_done(event => assert_equals(event.data, 'PASS'));
}, `Verifies that a form element's target can be a data url.`);
</script>
